changed back to integer to show working app

+++ db/rec_store.sql
@@ -11,8 +11,8 @@ CREATE TABLE albums(
   id SERIAL4 PRIMARY KEY,
   title VARCHAR(255),
   artist_id INT4 REFERENCES artists(id) ON DELETE CASCADE,
-  retail MONEY,
-  wholesale MONEY
+  retail int4,
+  wholesale int4
 );
 
 CREATE TABLE stocks(
